With dryrun.

+++ replace.go
@@ -2,14 +2,17 @@ package main
 
 import (
 	"flag"
+	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"regexp"
 	"strings"
+
+	"github.com/pmezard/go-difflib/difflib"
 )
 
-func replace(dir string, from *regexp.Regexp, to string) error {
+func replace(dir string, from *regexp.Regexp, to string, dryrun bool) error {
 	children, err := ioutil.ReadDir(dir)
 	if err != nil {
 		return err
@@ -20,7 +23,7 @@ func replace(dir string, from *regexp.Regexp, to string) error {
 		}
 		p := filepath.Join(dir, child.Name())
 		if child.IsDir() {
-			if err := replace(p, from, to); err != nil {
+			if err := replace(p, from, to, dryrun); err != nil {
 				return err
 			}
 		} else {
@@ -28,8 +31,20 @@ func replace(dir string, from *regexp.Regexp, to string) error {
 			if err != nil {
 				return err
 			}
-			if err := ioutil.WriteFile(p, from.ReplaceAll(b, []byte(to)), child.Mode()); err != nil {
-				return err
+			if dryrun {
+				diff := difflib.UnifiedDiff{
+					A:        difflib.SplitLines(string(b)),
+					B:        difflib.SplitLines(from.ReplaceAllString(string(b), to)),
+					FromFile: "Original",
+					ToFile:   "Replacement",
+					Context:  3,
+				}
+				text, _ := difflib.GetUnifiedDiffString(diff)
+				fmt.Printf("%v\t%v\n", p, text)
+			} else {
+				if err := ioutil.WriteFile(p, from.ReplaceAll(b, []byte(to)), child.Mode()); err != nil {
+					return err
+				}
 			}
 		}
 	}
@@ -39,6 +54,7 @@ func replace(dir string, from *regexp.Regexp, to string) error {
 func main() {
 	from := flag.String("from", "", "What to replace.")
 	to := flag.String("to", "", "What to replace it with.")
+	dryrun := flag.Bool("dryrun", false, "Don't perform any replacement, just output the ones that would have been made.")
 
 	flag.Parse()
 
@@ -56,7 +72,7 @@ func main() {
 	if err != nil {
 		panic(err)
 	}
-	if err := replace(cwd, fromReg, *to); err != nil {
+	if err := replace(cwd, fromReg, *to, *dryrun); err != nil {
 		panic(err)
 	}
 }
